Move partial date functionality into separate module
