+++ mep/accounts/tests/test_accounts_models.py
@@ -3,15 +3,13 @@ import re
 from unittest.mock import patch
 
 from dateutil.relativedelta import relativedelta
-from django.db import models
 from django.db.models.query import QuerySet
 from django.core.validators import ValidationError
 from django.test import TestCase
 import pytest
 
 from mep.accounts.models import Account, Address, \
-    Borrow, Event, Purchase, Reimbursement, Subscription, CurrencyMixin, \
-    DatePrecisionField, DatePrecision, PartialDate, PartialDateMixin
+    Borrow, Event, Purchase, Reimbursement, Subscription, CurrencyMixin
 from mep.books.models import Item
 from mep.people.models import Person, Location
 from mep.footnotes.models import Bibliography, SourceType
@@ -790,86 +788,6 @@ class TestBorrow(TestCase):
         assert borrow.item_status == borrow.ITEM_MISSING
 
 
-class TestPartialDateMixin(TestCase):
-
-    class PartialMixinObject(PartialDateMixin):
-
-        class Meta:
-            abstract = True
-
-    def test_calculate_date(self):
-
-        pmo = self.PartialMixinObject()
-        with pytest.raises(ValueError):
-            # unsupported date name should error
-            pmo.calculate_date('bogus')
-
-        # partial date
-        pmo.calculate_date('start_date', '1935-05')
-        assert pmo.start_date == datetime.date(1935, 5, 1)
-        assert pmo.start_date_precision.year
-        assert pmo.start_date_precision.month
-        assert not pmo.start_date_precision.day
-
-        # 1900 date = unknown by project convention
-        pmo.calculate_date('end_date', '1901-06-30')
-        assert pmo.end_date == datetime.date(1901, 6, 30)
-        assert not pmo.end_date_precision.year
-        assert pmo.end_date_precision.month
-        assert pmo.end_date_precision.day
-
-        # earliest/latest dates
-        early = datetime.date(1930, 11, 5)
-        late = datetime.date(1930, 11, 25)
-        pmo.calculate_date('start_date', earliest=early, latest=late)
-        # stored as earliest date
-        assert pmo.start_date == early
-        assert pmo.partial_start_date == '1930-11'
-        # in this case, all but day match
-        assert pmo.start_date_precision.year
-        assert pmo.start_date_precision.month
-        assert not pmo.start_date_precision.day
-
-        # only year overlaps
-        late = datetime.date(1930, 12, 25)
-        pmo.calculate_date('start_date', earliest=early, latest=late)
-        assert pmo.partial_start_date == '1930'
-        assert pmo.start_date_precision.year
-        assert not pmo.start_date_precision.month
-
-        # different year but same month/day
-        late = datetime.date(1932, 11, 5)
-        pmo.calculate_date('start_date', earliest=early, latest=late)
-        assert pmo.start_date == early
-        assert pmo.partial_start_date == '--11-05'
-        assert not pmo.start_date_precision.year
-        assert pmo.start_date_precision.month
-        assert pmo.start_date_precision.day
-
-        # no overlap?
-        late = datetime.date(1932, 12, 22)
-        pmo.calculate_date('start_date', earliest=early, latest=late)
-        assert not pmo.partial_start_date
-        assert not pmo.start_date_precision
-
-    def test_date_range(self):
-
-        # test both dates being the same returning date in partial date format
-        pmo = self.PartialMixinObject()
-        pmo.calculate_date('start_date', '1930-01-01')
-        pmo.calculate_date('end_date', '1930-01-01')
-        assert pmo.date_range == '1930-01-01'
-
-        # test that two dates produce / joined dates
-        pmo.calculate_date('end_date', '1930-01-02')
-        assert pmo.date_range == '1930-01-01/1930-01-02'
-
-        # test that an unknown date is rendered as ??
-        pmo.end_date = None
-        pmo.calculate_date('end_date')
-        assert pmo.date_range == '1930-01-01/??'
-
-
 class TestCurrencyMixin(TestCase):
 
     # create test currency model to test mixin behavior
@@ -898,74 +816,3 @@ class TestCurrencyMixin(TestCase):
         assert coin.currency_symbol() == 'foo'
 
 
-class TestPartialDates(TestCase):
-
-    # test object for partial date descriptor behavior
-    class PartialDateObject(models.Model):
-        date = None
-        partial_date = PartialDate('date', 'date_precision')
-        date_precision = DatePrecisionField()
-
-        class Meta:
-            abstract = True
-
-    # version that uses 1900 for unknown years
-    class PartialDateObject1900(PartialDateObject):
-        partial_date = PartialDate('date', 'date_precision', 1900)
-
-        class Meta:
-            abstract = True
-
-    def test_get(self):
-        pdo = self.PartialDateObject()
-        # should not error if date is not set
-        assert pdo.partial_date is None
-        # full precision
-        pdo.date = datetime.date(1901, 3, 5)
-        pdo.date_precision = DatePrecision.year | DatePrecision.month | DatePrecision.day
-        assert pdo.partial_date == '1901-03-05'
-        # partial precision
-        pdo.date_precision = DatePrecision.year | DatePrecision.month
-        assert pdo.partial_date == '1901-03'
-        pdo.date_precision = DatePrecision.month | DatePrecision.day
-        assert pdo.partial_date == '--03-05'
-        pdo.date_precision = DatePrecision.year
-        assert pdo.partial_date == '1901'
-        # change default unknown year value
-        pdo = self.PartialDateObject1900()
-        pdo.date = datetime.date(1900, 3, 5)
-        pdo.date_precision = DatePrecision.month | DatePrecision.day
-        assert pdo.partial_date == '--03-05'
-
-    def test_set(self):
-        pdo = self.PartialDateObject()
-        # full precision
-        pdo.partial_date = '1901-03-05'
-        assert pdo.date == datetime.date(1901, 3, 5)
-        assert pdo.date_precision == DatePrecision.year | DatePrecision.month | DatePrecision.day
-        # partial precision
-        pdo.partial_date = '1901-03'
-        assert pdo.date == datetime.date(1901, 3, 1)
-        assert pdo.date_precision == DatePrecision.year | DatePrecision.month
-        pdo.partial_date = '--03-05'
-        assert pdo.date == datetime.date(1, 3, 5)
-        assert pdo.date_precision == DatePrecision.month | DatePrecision.day
-        pdo.partial_date = '1901'
-        assert pdo.date == datetime.date(1901, 1, 1)
-        assert pdo.date_precision == DatePrecision.year
-        # invalid partial precision
-        with pytest.raises(ValidationError):
-            pdo.partial_date = '05'
-        with pytest.raises(ValidationError):
-            pdo.partial_date = '1901--05'
-        with pytest.raises(ValidationError):
-            pdo.partial_date = 'definitely_not_a_date'
-        # should clear the values if None is passed
-        pdo.partial_date = None
-        assert pdo.date is None
-        assert pdo.date_precision is None
-        # change default unknown year value
-        pdo = self.PartialDateObject1900()
-        pdo.partial_date = '--03-05'
-        assert pdo.date == datetime.date(1900, 3, 5)
-        assert pdo.date_precision == DatePrecision.month | DatePrecision.day
